Componentizado HourContainer
